requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
python-dateutil>=2.9
pydantic>=2.6
google-genai>=1.0.0
